refactor: use $DISPLAY for conditional termguicolors
note:
previously using `$TERM` to determine activating `termguicolors`,
now use `$DISPLAY` instead of `$TERM` because sometimes i use tmux
in tty and the `$TERM` value is not `linux` or (maybe i use) other
unix-like os such as bsd.

the downside of this method is that, this is probably only works in
x11, not sure about wayland.